# Optimize concurrent.futures→asyncio state transfer with atomic snapshot