--- a/Lib/test/test_concurrent_futures/test_future.py
+++ b/Lib/test/test_concurrent_futures/test_future.py
@@ -291,7 +291,7 @@ class FutureTests(BaseTestCase):
         self.assertFalse(cancelled)
         self.assertIsNone(result)
         self.assertIsNone(exception)
-        
+
         # Test with a finished future (successful result)
         f = Future()
         f.set_result(42)
@@ -300,7 +300,7 @@ class FutureTests(BaseTestCase):
         self.assertFalse(cancelled)
         self.assertEqual(result, 42)
         self.assertIsNone(exception)
-        
+
         # Test with a finished future (exception)
         f = Future()
         exc = ValueError("test error")
@@ -310,7 +310,7 @@ class FutureTests(BaseTestCase):
         self.assertFalse(cancelled)
         self.assertIsNone(result)
         self.assertEqual(exception, exc)
-        
+
         # Test with a cancelled future
         f = Future()
         f.cancel()
@@ -319,26 +319,26 @@ class FutureTests(BaseTestCase):
         self.assertTrue(cancelled)
         self.assertIsNone(result)
         self.assertIsNone(exception)
-        
+
         # Test concurrent access (basic thread safety check)
         f = Future()
         f.set_result(100)
         results = []
-        
+
         def get_snapshot():
             for _ in range(1000):
                 snapshot = f._get_snapshot()
                 results.append(snapshot)
-        
+
         threads = []
         for _ in range(4):
             t = threading.Thread(target=get_snapshot)
             threads.append(t)
             t.start()
-        
+
         for t in threads:
             t.join()
-        
+
         # All snapshots should be identical for a finished future
         expected = (True, False, 100, None)
         for result in results: